refactor rust code out
